# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Bonza Bike Tours REST API",
		"description": "REST API description for the Bonza Bike Tours BMS dashboard.",
		"contact": {
			"name": "API Support",
			"email": "alexander.shumilov@toptal.com"
		},
		"version": "1.2"
	},
	"servers": [{
			"url": "bonzabiketours-api.herokuapp.com/api/v1",
			"description": "Heroku server - CI"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3001/api/v1",
			"description": "Staging server"
		},
		{
			"url": "bms.bonzabiketours.alexshumilov.ru:3002/api/v1",
			"description": "Development server"
		}
	],
	"security": [{
		"api_key": []
	}],
...
	"paths": {
...
		"/products": {
			"get": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Return products according to filters",
				"parameters": [{
						"name": "sortby",
						"description": "Search results order",
						"in": "query",
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "typeCodes",
						"description": "Specified type codes",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"$ref": "#/components/parameters/limit"
					},
					{
						"$ref": "#/components/parameters/page"
					}
				],
				"responses": {
					"200": {
						"description": "Products search result",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Create product",
				"requestBody": {
					"description": "Product description",
					"content": {
						"application/json": {
							"schema": {
								"title": "",
								"properties": {
									"name": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Created product identifier",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APICreatedEntityResponse"
								}
							}
						}
					}
				}
			}
		},
		"/products/order": {
			"put": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Updates sorting order of products",
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					}
				}
			}
		},
		"/products/types": {
			"get": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Returns product types",
				"responses": {
					"200": {
						"description": "Product types",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/ProductType"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/products/{productId}": {
			"get": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Return product",
				"parameters": [{
					"name": "productId",
					"description": "Product identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Product",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Product"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"put": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Update product",
				"parameters": [{
					"name": "productId",
					"description": "Product identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"requestBody": {
					"description": "Product photos",
					"content": {
						"multipart/form-data": {
							"schema": {
								"title": "",
								"properties": {}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Delete product",
				"parameters": [{
					"name": "productId",
					"description": "Product identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/responses/APISuccessResponse"
									}
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/products/{productId}/tourPricing": {
			"get": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Return tour pricing for product",
				"parameters": [{
						"name": "productId",
						"description": "Product identifier",
						"in": "path",
						"schema": {
							"type": "integer"
						},
						"required": true
					},
					{
						"name": "date",
						"description": "Requested date",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"required": true
					},
					{
						"name": "bookingPartnerId",
						"description": "Booking partner identifier",
						"in": "query",
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Tour pricing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/products/{productId}/miscPricing": {
			"get": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Return miscellaneous pricing for product",
				"parameters": [{
					"name": "productId",
					"description": "Product identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Misc pricing",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
		"/products/{productId}/seasons": {
			"get": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Returns price seasons for specific product",
				"parameters": [{
					"name": "productId",
					"description": "Product identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Price seasons list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			},
			"put": {
				"tags": [
					"products"
				],
				"security": [{
					"api_key": []
				}],
				"summary": "Updates price seasons for specific product",
				"parameters": [{
					"name": "productId",
					"description": "Product identifier",
					"in": "path",
					"schema": {
						"type": "integer"
					},
					"required": true
				}],
				"responses": {
					"200": {
						"description": "Operation status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/responses/APISuccessResponse"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/APIUnauthorizedResponse"
					},
					"404": {
						"$ref": "#/components/responses/APINotFoundResponse"
					}
				}
			}
		},
...
	},
	"components": {
		"responses": {
...
			"APISuccessResponse": {
				"description": "Operation was completed successfully",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"message": {
									"type": "string"
								}
							}
						}
					}
				}
			},
...
			"APIUnauthorizedResponse": {
				"description": "Authorize in order to access the resource",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APINotFoundResponse": {
				"description": "Requested resource was not found",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"status": {
									"type": "string"
								},
								"errorType": {
									"type": "string"
								},
								"errorMessage": {
									"type": "string"
								}
							}
						}
					}
				}
			},
			"APICreatedEntityResponse": {
				"description": "Returning identifier of created entity",
				"content": {
					"application/json": {
						"schema": {
							"properties": {
								"id": {
									"type": "integer"
								}
							}
						}
					}
				}
			}
		},
		"parameters": {
...
			"limit": {
				"name": "limit",
				"in": "query",
				"description": "Specifies maximum number of fetched items",
				"schema": {
					"type": "integer"
				}
			},
			"page": {
				"name": "page",
				"in": "query",
				"description": "Specifies page of fetched items (offsetting)",
				"schema": {
					"type": "integer"
				}
			}
		},
		"schemas": {
...
			"ProductType": {
				"properties": {
					"typeId": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					}
				}
			},
			"Product": {
				"properties": {
					"productId": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					}
				}
			},
...
		}
	}
}